AI_Project/AI_Code_Guardian_Agent/requirements.txt
fpdf
streamlit
google-generativeai